fix: receive ticket
