+++ src/hooks/useTickets/index.ts
@@ -6,18 +6,13 @@ import {
   DONATION_TOAST_SEEN_AT_KEY,
 } from "lib/localStorage/constants";
 
-import { useAuthentication } from "contexts/authenticationContext";
-import { getLocalStorageItem, setLocalStorageItem } from "@ribon.io/shared/lib";
+import { getLocalStorageItem } from "@ribon.io/shared/lib";
 import { useIntegrationContext } from "contexts/integrationContext";
 import { PLATFORM } from "utils/constants/Application";
-import { showToast } from "lib/Toast";
-import { theme } from "@ribon.io/shared/styles";
 import { todayDate } from "lib/dateUtils";
-import { useTranslation } from "react-i18next";
 
 export function useTickets() {
   const { currentUser } = useCurrentUser();
-  const { isAuthenticated } = useAuthentication();
   const {
     canCollectByExternalIds,
     canCollectByIntegration,
@@ -27,9 +22,6 @@ export function useTickets() {
 
   const { currentIntegrationId, externalId } = useIntegrationContext();
   const externalIds = externalId?.split(",");
-  const { t } = useTranslation("translation", {
-    keyPrefix: "donations.causesScreen.ticketSection",
-  });
 
   async function hasReceivedTicketToday() {
     const donationToastSeenAtKey = await getLocalStorageItem(
@@ -51,7 +43,9 @@ export function useTickets() {
 
   async function handleCanCollect() {
     if (externalIds && externalIds.length > 0) {
+      console.log(externalIds);
       const { canCollect } = await canCollectByExternalIds(externalIds);
+      console.log("canCollect", canCollect);
       return canCollect;
     } else if (currentIntegrationId) {
       const { canCollect } = await canCollectByIntegration(
@@ -65,6 +59,7 @@ export function useTickets() {
   }
 
   async function handleCollect() {
+    console.log("collect");
     if (externalIds && externalIds.length > 0 && currentIntegrationId) {
       await collectByExternalIds(
         externalIds,
@@ -81,40 +76,9 @@ export function useTickets() {
     }
   }
 
-  async function receiveTicket() {
-    const canCollect = await handleCanCollect();
-
-    if (canCollect && !hasReceivedTicketToday()) {
-      if (isAuthenticated()) {
-        await handleCollect();
-      }
-      await setLocalStorageItem(
-        DONATION_TOAST_SEEN_AT_KEY,
-        Date.now().toString(),
-      );
-      await setLocalStorageItem(
-        DONATION_TOAST_INTEGRATION,
-        currentIntegrationId?.toLocaleString(),
-      );
-
-      showToast({
-        type: "custom",
-        message: t("ticketToast"),
-        position: "bottom",
-        navigate: "GiveTicketScreen",
-        icon: "confirmation_number",
-        backgroundColor: theme.colors.brand.primary[50],
-        iconColor: theme.colors.brand.primary[600],
-        borderColor: theme.colors.brand.primary[600],
-        textColor: theme.colors.brand.primary[600],
-      });
-    }
-  }
-
   return {
     handleCanCollect,
     handleCollect,
-    receiveTicket,
     hasReceivedTicketToday,
   };
 }
